Remove lex, format, and minify from README
There's not a lot of reason to include `lex`, `format`, and `minify` in the README. They just clutter up the docs, create confusion about the point of crossplane, and keeping them in creates a responsibility to maintain them. I don't think we should remove the capabilities from the tool, though, to maintain backwards compatibility.