Make `http` `types` directives analysable
The syntax of `http` `types` blocks is distinct, in that their child
directives are arbitrary MIME types - they therefore can't be analysed
when `strict` or `check_ctx` are enabled, and no meaningful analysis can
be performed when `check_args` is enabled because the arity of child
directives is unknown. Skip the checks performed when `strict` and
`check_ctx` are enabled, and hint to the analyser that all directives
inside a `types` block accept one or more arguments so that a meaningful
check can be performed when `check_args` is enabled.

Fixes #101.